break cgi-reports into multiple modules
